+++ src/components/NavbarItem.jsx
@@ -1,12 +1,9 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 import style from '../style/Navbar.module.css';
+import { getTitle } from '../js/routes_names';
 
 class NavbarItem extends React.Component {
-  handleChange = () => {
-    // this.props.showNavbar();
-  };
-
   render() {
     return (
       <div className={style.navbarItem}>
@@ -21,7 +18,7 @@ class NavbarItem extends React.Component {
             className={style.image_navbar_item}
             alt={this.props.title}
           />
-          <h4>{this.props.title}</h4>
+          <h4>{getTitle(this.props.title)}</h4>
         </NavLink>
       </div>
     );
