+++ src/components/tabs/Calendar.jsx
@@ -1,10 +1,6 @@
 import React from 'react';
 
 class Calendar extends React.Component {
-  // constructor(props) {
-  //   super(props);
-  // }
-
   render() {
     return <div>Calendar</div>;
   }
