+++ src/components/tabs/Home.jsx
@@ -2,10 +2,6 @@ import React from 'react';
 import style from '../../style/Home.module.css';
 
 class Home extends React.Component {
-  // constructor(props) {
-  //   super(props);
-  // }
-
   render() {
     return (
       <div className={style.home_content}>
