+++ src/components/tabs/Settings.jsx
@@ -1,10 +1,6 @@
 import React from 'react';
 
 class Settings extends React.Component {
-  // constructor(props) {
-  //   super(props);
-  // }
-
   render() {
     return <div>Settings</div>;
   }
