+++ src/components/tabs/Statistics.jsx
@@ -1,10 +1,6 @@
 import React from 'react';
 
 class Statistics extends React.Component {
-  // constructor(props) {
-  //   super(props);
-  // }
-
   render() {
     return <div>Statistics</div>;
   }
