+++ src/components/tabs/Tasks/Button.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import style_button from '../../../style/Button.module.css';
+import styleButton from '../../../style/Button.module.css';
 
 class Button extends React.Component {
   constructor(props) {
@@ -13,15 +13,9 @@ class Button extends React.Component {
 
   render() {
     return (
-      <div
-        className={style_button.button}
-        onClick={this.changePost}
-        role='button'
-        onKeyPress={() => {}}
-        tabIndex='0'
-      >
+      <button className={styleButton.button} onClick={this.changePost}>
         {this.props.title}
-      </div>
+      </button>
     );
   }
 }
