+++ src/components/tabs/Tasks/Palette.jsx
@@ -24,14 +24,11 @@ class Palette extends React.Component {
 
   render() {
     return (
-      <div
+      <button
         onClick={this.chooseBackgroundColor}
         className={style.paletteItem}
         style={this.setBackgroundColor()}
-        role='button'
-        onKeyPress={() => {}}
-        tabIndex='0'
-      ></div>
+      ></button>
     );
   }
 }
