+++ src/components/tabs/Tracks/Tracks.jsx
@@ -1,13 +1,82 @@
 import React from 'react';
 import Track from './Track';
 import style from '../../../style/Tracks.module.css';
+import {
+  setDataInStorage,
+  deleteDataFromStorage,
+} from '../../../js/setDataInStorage';
+import getId from '../../../js/getId';
+import { getTracks } from '../../../js/tracks';
 
 class Tracks extends React.Component {
-  render() {
-    let tracks;
+  constructor(props) {
+    super(props);
+    this.state = {
+      newTrack: '',
+      newTrackName: '',
+    };
 
-    if (this.props.state === undefined) tracks = this.props.tracks;
-    else tracks = this.props.state.tracks;
+    this.changeTrack = this.changeTrack.bind(this);
+  }
+
+  componentWillMount() {
+    const tracks = getTracks(this.props.email);
+    this.setState({ tracks });
+  }
+
+  addNewTrack = (projectId, projectTitle) => {
+    const { newTrack } = this.state;
+    const track = {
+      taskId: projectId,
+      userId: this.props.email,
+      trackId: getId(),
+      taskName: projectTitle,
+      trackNode: newTrack,
+      trackDate: new Date(),
+    };
+
+    const { tracks } = this.state;
+    this.setState({
+      tracks: [...tracks, track],
+    });
+    this.clearInputFields();
+    this.setState({
+      tasks: this.state.tasks.map((task) => {
+        if (task.taskId === projectId) {
+          return { ...task, showTrackField: false };
+        }
+        return task;
+      }),
+    });
+
+    setDataInStorage(track, 'trac');
+  };
+
+  deleteTrack(id) {
+    const { tracks } = this.state;
+    this.setState({ tracks: tracks.filter((track) => track.trackId !== id) });
+    deleteDataFromStorage(id, 'trac');
+  }
+
+  changeTrack(id, action) {
+    switch (action) {
+      case 'delete': {
+        this.deleteTrack(id);
+        break;
+      }
+      case 'edit': {
+        alert('edit');
+        break;
+      }
+      default:
+        console.log(id);
+    }
+  }
+  render() {
+    let tracks = this.state.tracks;
+
+    if (this.props.tracks === undefined) tracks = this.state.tracks;
+    else tracks = this.props.tracks;
 
     const tracksCount = tracks.length;
 
@@ -30,19 +99,19 @@ class Tracks extends React.Component {
           {tracks.map((track, index) => (
             <Track
               track={track}
-              changeTrack={this.props.changeTrack}
+              changeTrack={this.changeTrack}
               index={index}
               key={track.id}
             />
           ))}
         </div>
-        <div className={style.page_numbers}>
+        {/* <div className={style.page_numbers}>
           <div className={style.page_number}>1</div>
           <div className={style.page_number}>2</div>
           <div className={style.page_number}>3</div>
           <div className={style.page_number}>4</div>
           <div className={style.page_number}>5</div>
-        </div>
+        </div> */}
       </div>
     );
   }
