+++ src/components/tabs/Users/Users.jsx
@@ -3,27 +3,150 @@ import style from '../../../style/Users.module.css';
 import User from './User';
 import Popup from 'reactjs-popup';
 import PopUp from './PopUp';
+import {
+  deleteDataFromStorage1,
+  setDataInStorage,
+} from '../../../js/setDataInStorage';
+import { getUsers } from '../../../js/users';
+import { ADMIN } from '../../../js/roles';
 
 class Users extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      startDate: '',
+      sex: 'Male',
+      direction: 'Member',
+      date: new Date(),
+      name: '',
+      lastName: '',
+      email: '',
+      education: '',
+      age: '',
+      university: '',
+      mathScore: '',
+      address: '',
+      mobilePhone: '',
+      skype: '',
+    };
+
+    this.changeUser = this.changeUser.bind(this);
+  }
+
+  componentWillMount() {
+    const users = getUsers();
+    this.setState({ users });
+  }
+
+  handleInputChange = (name, event) => {
+    this.setState({
+      [name]: event.target.value,
+    });
+  };
+
+  dateOfBirth = (date) => this.setState({ date });
+
+  onChangeStartDate = (startDate) => this.setState({ startDate });
+
+  onChangeDirection = (event) =>
+    this.setState({ direction: event.target.value });
+
+  onChangeSex = (event) => this.setState({ sex: event.target.value });
+
+  getCurrentState = () => {
+    return [
+      this.state.name,
+      this.state.lastName,
+      this.state.email,
+      this.state.education,
+      this.state.age,
+      this.state.university,
+      this.state.mathScore,
+      this.state.address,
+      this.state.mobilePhone,
+      this.state.skype,
+    ];
+  };
+
+  deleteUser(id) {
+    const { users } = this.state;
+    this.setState({ users: users.filter((user) => user.email !== id) });
+    deleteDataFromStorage1(id, 'user');
+  }
+
+  changeUser(id, action) {
+    switch (action) {
+      case 'delete': {
+        this.deleteUser(id);
+        break;
+      }
+      case 'edit': {
+        alert('edit');
+        break;
+      }
+      default:
+        console.log(id);
+    }
+  }
+
+  addNewUser = () => {
+    const {
+      sex,
+      direction,
+      date,
+      startDate,
+      name,
+      lastName,
+      email,
+      education,
+      age,
+      university,
+      mathScore,
+      address,
+      mobilePhone,
+      skype,
+    } = this.state;
+    const newUser = {
+      sex,
+      direction,
+      date,
+      startDate,
+      name,
+      lastName,
+      email,
+      education,
+      age,
+      university,
+      mathScore,
+      address,
+      mobilePhone,
+      skype,
+    };
+
+    const { users } = this.state;
+    this.setState({
+      users: [...users, newUser],
+    });
+
+    setDataInStorage(newUser, 'user');
+  };
   render() {
-    console.log('his.props');
-    console.log(this.props);
-    const users = this.props.state.users;
+    const { users } = this.state;
     const usersCount = users.length;
     return (
       <div className={style.users_container}>
-        {this.props.direction === 'Admin' ? (
+        {this.props.direction === ADMIN ? (
           <Popup modal trigger={<button>Add user</button>}>
             {(close) => (
               <PopUp
                 close={close}
-                state={this.props.state}
-                handleInputChange={this.props.handleInputChange}
-                addNewUser={this.props.addNewUser}
-                dateOfBirth={this.props.dateOfBirth}
-                onChangeStartDate={this.props.onChangeStartDate}
-                onChangeDirection={this.props.onChangeDirection}
-                getCurrentState={this.props.getCurrentState}
+                state={this.state}
+                handleInputChange={this.handleInputChange}
+                addNewUser={this.addNewUser}
+                dateOfBirth={this.dateOfBirth}
+                onChangeStartDate={this.onChangeStartDate}
+                onChangeDirection={this.onChangeDirection}
+                getCurrentState={this.getCurrentState}
               />
             )}
           </Popup>
@@ -48,18 +171,18 @@ class Users extends React.Component {
             <User
               user={user}
               key={user.id}
-              changeUser={this.props.changeUser}
+              changeUser={this.changeUser}
               direction={this.props.direction}
             />
           ))}
         </div>
-        <div className={style.page_numbers}>
+        {/* <div className={style.page_numbers}>
           <div className={style.page_number}>1</div>
           <div className={style.page_number}>2</div>
           <div className={style.page_number}>3</div>
           <div className={style.page_number}>4</div>
           <div className={style.page_number}>5</div>
-        </div>
+        </div> */}
       </div>
     );
   }
