+++ src/components/tabs/Workflow.jsx
@@ -1,32 +1,83 @@
 import React from 'react';
 import style from '../../style/Workflow.module.css';
 import Task from './Task';
+import getProjects from '../../js/projects';
+import { setChangeDataInStorage } from '../../js/setDataInStorage';
 
 class Workflow extends React.Component {
-  // constructor(props) {
-  //   super(props);
-  // }
+  constructor(props) {
+    super(props);
 
-  getToDoTask(tasks) {
-    const newTasks = tasks.filter((task) => task.state === 'toDo');
+    this.changePost = this.changePost.bind(this);
+  }
+  componentWillMount() {
+    const tasks = getProjects(this.props.email);
+    this.setState({ tasks });
+  }
+
+  getTask(tasks, name) {
+    const newTasks = tasks.filter((task) => task.state === name);
     return newTasks;
   }
 
-  getInProgressTask(tasks) {
-    const newTasks = tasks.filter((task) => task.state === 'inProgress');
-    return newTasks;
+  handleInputChange = (name, event) => {
+    this.setState({
+      [name]: event.target.value,
+    });
+  };
+
+  changeState(task, newState) {
+    return { ...task, state: newState, showEditFields: false };
   }
 
-  getCompletedTask(tasks) {
-    const newTasks = tasks.filter((task) => task.state === 'completed');
-    return newTasks;
+  showChangeTaskField(task, name) {
+    if (task[name]) {
+      return { ...task, [name]: false };
+    }
+    return { ...task, [name]: true };
+  }
+
+  changePost(id, action, color) {
+    this.setState({
+      tasks: this.state.tasks.map((task) => {
+        if (task.taskId === id) {
+          switch (action) {
+            case `changeStatus`: {
+              task = this.changeStatus(task);
+              break;
+            }
+            case `toDo`: {
+              task = this.changeState(task, `toDo`);
+              break;
+            }
+            case `inProgress`: {
+              task = this.changeState(task, `inProgress`);
+              break;
+            }
+            case `completed`: {
+              task = this.changeState(task, `completed`);
+              break;
+            }
+            case `showEditFields`: {
+              task = this.showChangeTaskField(task, 'showEditFields');
+              break;
+            }
+            default:
+              return task;
+          }
+          setChangeDataInStorage(task, 'taskId', 'task');
+        }
+
+        return task;
+      }),
+    });
   }
 
   render() {
-    const tasks = this.props.state.tasks;
-    const toDoTask = this.getToDoTask(tasks);
-    const inProgressTask = this.getInProgressTask(tasks);
-    const completedTask = this.getCompletedTask(tasks);
+    const tasks = this.state.tasks;
+    const toDoTask = this.getTask(tasks, 'toDo');
+    const inProgressTask = this.getTask(tasks, 'inProgress');
+    const completedTask = this.getTask(tasks, 'completed');
 
     return (
       <div className={style.workflow_container}>
@@ -35,11 +86,7 @@ class Workflow extends React.Component {
             ToDo <span className={style.count_task}>({toDoTask.length})</span>
           </h1>
           {toDoTask.map((task) => (
-            <Task
-              task={task}
-              key={task.id}
-              changePost={this.props.changePost}
-            />
+            <Task task={task} key={task.id} changePost={this.changePost} />
           ))}
         </section>
         <section className={style.inProgress_container}>
@@ -48,11 +95,7 @@ class Workflow extends React.Component {
             <span className={style.count_task}>({inProgressTask.length})</span>
           </h1>
           {inProgressTask.map((task) => (
-            <Task
-              task={task}
-              key={task.id}
-              changePost={this.props.changePost}
-            />
+            <Task task={task} key={task.id} changePost={this.changePost} />
           ))}
         </section>
         <section className={style.completed_container}>
@@ -61,11 +104,7 @@ class Workflow extends React.Component {
             <span className={style.count_task}>({completedTask.length})</span>
           </h1>
           {completedTask.map((task) => (
-            <Task
-              task={task}
-              key={task.id}
-              changePost={this.props.changePost}
-            />
+            <Task task={task} key={task.id} changePost={this.changePost} />
           ))}
         </section>
       </div>
