+++ src/js/checkTimeOfTask.js
@@ -2,7 +2,7 @@ const getDeadlineStatus = (deadlineDate) => {
   const currentTime = new Date().getTime();
   let timeToDeadline = (deadlineDate.getTime() - currentTime) / 1000 / 60 / 60; // time to deadline in hours
   if (timeToDeadline > 0) return 1;
-  else return 0;
+  return 0;
 };
 
 const getTime = (deadlineDate) => {
@@ -14,12 +14,16 @@ const getTime = (deadlineDate) => {
     str = 'delays';
     timeToDeadline = timeToDeadline * -1;
   }
+
   if (timeToDeadline < 24) return `${Math.floor(timeToDeadline)} hours ${str}`;
   timeToDeadline = timeToDeadline / 24;
+
   if (timeToDeadline < 30) return `${Math.floor(timeToDeadline)} days ${str}`;
   timeToDeadline = timeToDeadline / 30;
+
   if (timeToDeadline < 12) return `${Math.floor(timeToDeadline)} months ${str}`;
   timeToDeadline = timeToDeadline / 12;
+
   return `${Math.floor(timeToDeadline)} years ${str}`;
 };
