+++ src/js/getId.js
@@ -1,0 +1,5 @@
+const getId = () => {
+  return String(Math.round(Math.random() * (100000000 + 10000000) - 10000000));
+};
+
+export default getId;
