+++ src/js/roles.js
@@ -1,0 +1,5 @@
+const ADMIN = 'Admin';
+const MENTOR = 'Mentor';
+const USER = 'Member';
+
+export { ADMIN, MENTOR, USER };
